+++ MapView/Map/RMDBMapSource.m
@@ -160,7 +160,7 @@
     _center.latitude = [self getPreferenceAsFloat:kCoverageCenterLatitudeKey];
     _center.longitude = [self getPreferenceAsFloat:kCoverageCenterLongitudeKey];
 
-    RMLog(@"Tile size: %d pixel", self.tileSideLength);
+    RMLog(@"Tile size: %lu pixel", (unsigned long)self.tileSideLength);
     RMLog(@"Supported zoom range: %.0f - %.0f", self.minZoom, self.maxZoom);
     RMLog(@"Coverage area: (%2.6f,%2.6f) x (%2.6f,%2.6f)",
           _topLeft.latitude,
